requirements.txt
ai2thor>=0+8524eadda94df0ab2dbb2ef5a577e4d37c712897
boto3>=1.42.75
botocore>=1.42.75
colorama>=0.4.6
compress_json>=1.1.1
compress_pickle>=2.1.0
fastapi>=0.135.2
langchain_core>=1.2.22
langchain_openai>=1.1.12
matplotlib>=3.10.8
moviepy>=2.2.1
numpy>=2.4.3
openai>=2.29.0
Pillow>=12.1.1
procthor>=0.0.1.dev2
pydantic>=2.12.5
pydantic_settings>=2.13.1
python-dotenv>=1.2.2
rtree>=1.4.1
scipy>=1.17.1
sentence_transformers>=5.2.2
Shapely>=2.1.2
SQLAlchemy>=2.0.48
torch>=2.10.0
tqdm>=4.67.2
transformers>=5.0.0
trimesh>=4.11.1
